dev: moved store address tour to guided-tour folder
